VisitVerse_face/ivms_backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==9.5.0
deepface==0.0.93
opencv-python-headless==4.9.0.80
tf-keras>=2.16
